# swagger.yaml
openapi: 3.0.0
info:
  title: D-Link - API - Product Images
  description: |
    This API primarily provides a facility to retrieve 
    image urls associated with a D-Link product code.
    A product code may have none, one or many images 
    available. Therefore image url retrieval can optionally be 
    filtered on keywords such as 'left', 'right' or a combination e.g. 
    'front middle left'. 
servers:
  - url: http://jdocker01.dlinkjde.int:8080/api/v1/productimages
paths:
  /ping:
    get: 
      summary: | 
        Pings this API for a simple OK response.
      description: |
        Simple API call, no parameters, quickly check your code 
        can connect with this API and get a response back.
      produces:
      - "application/json"
      responses:
        '200':    # status code
          description: A JSON object with success and a mesage with 'OK'.

  /products:
    get:
      summary: |
        Returns a distinct list of all Products which have at least one image url available.
      description: | 
        Returns a distinct list of all product codes stored in this 
        system where a product code has at least one image url available.
      produces:
      - "application/json"
      responses:
        '200':    # status code
          description: A JSON array of product codes

  /images:
    post:
      summary: Add a new Image entry with Product code and defining meta keywords such as 'left' or 'front left'.
      description: |
        Pass a Product code, image url and one or more meta keyword tags e.g. left or front to describe the image.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                product:          	# <!--- form field name
                  type: string
                image:   			# <!--- form field name
                  type: string
                meta:    		# <!--- form field name
                  type: string
              required:
                - product
                - image      
                - meta      
      responses:
        '201':
          description: "Created"  
        '409':   
          description: "Duplicate key value - Entry already exists."  
        '422':   
          description: "Missing parameter value etc."  
          description: A JSON array of image urls

    get:
      summary: Returns a list of image urls available for a Product code.
      description: |
        Pass a Product code for a list of all available image urls 
        stored in the system.
        Optionally filter this list by providing one or word filter words
        for example 'left' or 'front left' 
      parameters:
      - in: query
        name: product
        schema:
          type: string
        required: true
        description: Product code of image urls to get
      - in: query
        name: filter
        schema:
          type: string
        required: false
        description: Optional meta keyword filters to restrict results e.g. 'Left' or 'bak right'
      responses:
        '200':    # status code
          description: A JSON array of image urls

  /images/{id}:
    put:
      summary: Updates an image entry by Id.
      description: |
        Pass a Product code for a list of all available image urls 
        stored in the system.
        Optionally filter this list by providing one or word filter words
        for example 'left' or 'front left' 
      parameters:
      - in: path
        name: id
        description: Id of product/image entry you want to update
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                product:          	# <!--- form field name
                  type: string
                image:   			# <!--- form field name
                  type: string
                meta:    		# <!--- form field name
                  type: string
              required:
                - product
                - image      
                - meta      
      responses:
        '204':    # status code
          description: Update confirmed - no content returned
        '422':
          description: Error with parameters values see error messages in response 

    delete:
      summary: Deletes an image entry by Id.
      description: |
        Pass a Product code for a list of all available image urls 
        stored in the system.
        Optionally filter this list by providing one or word filter words
        for example 'left' or 'front left' 
      parameters:
      - in: path
        name: id
        description: Id of product/image entry you want to delete
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Delete confirmed - no content returned
        '422':
          description: Error with parameters values see error messages in response 
definitions:
  Images:
    type: "Object"
    properties:
      id:
        typs: "integer"
        format: "int64"
      product:
        type: "string"
      url:
        type: "string"
      meta:
        type: "string"
    json:
      name: "Images"